app/requirements.txt
openai
qdrant-client
fastembed
python-dotenv
